#10 New OpenShiftStrategy for Passport Added
This new authentication strategy lives alongside the existing Auth0Strategy.  Which one is the live strategy in use is now controlled via the OAUTH2_PROVIDER environment variable.  At the same time, we've dropped the _AUTH0_ prefixes to OAuth environment variables, as it's growing into a more generic OAuth2 strategy.

Please note that we have breaking changes to the 'user' table.  It needs to be dropped and re-created in live instances already using it.